refactor: move Lotto.kt to a higher-level package
