+++ src/main/kotlin/lotto/domain/LottoTicketGenerator.kt
@@ -1,6 +1,7 @@
 package lotto.domain
 
 import camp.nextstep.edu.missionutils.Randoms
+import lotto.Lotto
 
 object LottoTicketGenerator {
     fun generate(): Lotto {
