+++ src/test/kotlin/lotto/domain/WinningNumbersTest.kt
@@ -1,6 +1,7 @@
 package lotto.domain
 
 import WinningNumbers
+import lotto.Lotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
